# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warung Buah API",
        "description": "API Documentation untuk aplikasi Warung Buah - Sistem manajemen transaksi buah dengan fitur lengkap untuk supplier, produk, kategori, transaksi, dan pembayaran.",
        "contact": {
            "name": "API Support",
            "email": "support@warungbuah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pelanggan": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "List pelanggan",
                "operationId": "9eca3b486cba914398aee5ddaaac6162",
                "parameters": [
                    {
                        "name": "cari",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "aktif",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Tambah pelanggan baru",
                "operationId": "0b7e2a57081d435bcd6d94c6b1fe6c53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "telepon": {
                                        "type": "string"
                                    },
                                    "toko": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "catatan": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pelanggan/all": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Semua pelanggan aktif (untuk dropdown)",
                "operationId": "1829281bcf04cffbacbc04ee9dc3eb48",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pelanggan/{id}": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Detail pelanggan + ringkasan piutang & deposit",
                "operationId": "fff2dafa2fd882e967f08c48758150ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Update pelanggan",
                "operationId": "08cd65276654dd7e23c413e4e3dd49fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Hapus pelanggan",
                "operationId": "df4d852b4c05bae6680fa1dd3183e9ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}